Implement parallel median calculation
Median calculation for a list of images in the parmed function was not
implemented in parallel because the original authors couldn't figure out
why it wasn't working. We did, and we fixed it.

The problem with the previous implementation (that was commented) was
two-fold: first, the outer 2 loops that were under parallelisation (with
an "omp parallel for collapse(2)" clause) corresponded to the outer two
dimensions of the list of images (depth and spectrum). However not all
images (certainly not ours) have these two extra dimensions, so it makes
sense to parallelise over the lower two dimensions (width and height)
instead. Once this problem was overcome, the second problem arose: the
code used Rcpp's median function, which internally allocates new Rcpp
vectors, which in turn require calling some of the R API that isn't
thread-safe.

This commit fixes the problems in one go. First, it parallelises the
computation of medians across width instead of depth and spectrum, which
maps better to common use cases (images usually have more pixels than
CPUs in the host machine, so there's no need to parallelise on height as
well). Secondly, instead of using Rcpp's data structures and median
function, it uses a collection of pre-allocated std::vectors (one per
thread) to hold the data that will be median'd, then it performs its own
median calculation using nth_element + max_element (similar to how
everyone, including Rcpp and Cimg, does).  Particular care is taken to
cater for the rm_na option: when set, we only copy the elements that are
not NA into the pre-allocated vectors, and perform the median on those.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>