Use correct OpenMP macro checks
The cimg_use_openmp macro is *always* defined by CImg.h, but its *value*
is either 0 or 1 depending on whether OpenMP is defined or not. This was
breaking compilation on platforms without OpenMP.

This also uncovered an issue with our code calling Rcpp::cerr instead of
Rcpp::Rcerr, which has been fixed now.

This fix should address #5.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>