Fix median result when all values are NAs
Our new implementation of imager::parmed was not taking into account the
case when it could be invoked with na.rm=T, and all the input pixels are
NAs. In such a case the vector with actual values to be median'd will be
empty, which isn't handled correctly by the rest of the _get_median
function.

The solution is to check for this case, and handle it separately by
simply returning NA.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>